Add files via upload
Online portal system is developed using JSP Server as
front end and MySQL as back end. The aim of the project is to provide an interface to school
or college maintenance. The Student result analysis and Performance reporting system
provides an interface to school or college maintenance. It can be used by educational
institutes or coaching classes to investigate student performance easily. It is used in
collecting the performance report of students according to year, branch, section and subject
wise which is a quite difficult and time consuming process in every college. It deals with all
kinds of student details, course details, class details, and academic-related marks details and
reports.